refactor: change the path of the routes file && update useNavigation hook
