+++ src/components/Card/index.tsx
@@ -1,7 +1,6 @@
-import { useNavigation } from '@react-navigation/native';
 import React, { FC, useCallback } from 'react';
 import { ImageSourcePropType } from 'react-native';
-import { routes } from 'src/constants/routes';
+import { routes, useNavigation } from 'src/navigation';
 import { CardBox, CardImg, CardItem, CardStatus, CardTitle } from './styled';
 
 export interface CardProps {
@@ -12,7 +11,7 @@ export interface CardProps {
 }
 
 export const Card: FC<CardProps> = ({ id, name, status, image }) => {
-  const navigation = useNavigation<any>();
+  const navigation = useNavigation();
 
   const handleNavigateToScreen = useCallback(() => {
     navigation.navigate(routes.CharacterStackScreen, {
